refactor(library): divide services.py to loan_service.py and reservation_service.py
